selfwork 1-n

+++ app/Http/Controllers/ServiziController.php
@@ -6,6 +6,7 @@ use App\Http\Controllers\Controller;
 use App\Http\Requests\ServiceRequest;
 use App\Models\Service;
 use Illuminate\Http\Requests;
+use Illuminate\Support\Facades\Auth;
 
 
 class ServiziController extends Controller
@@ -28,6 +29,7 @@ public function create() {
         $servizio->description = $request->description;
         $servizio->price = $request->price;
         $servizio->img = $request->file('img')->store('images', 'public');
+        $servizio->user_id = Auth::user()->id;
        $servizio->save();
 
        return redirect()->route('home')->with('successMessage','Servizio creato con successo!');
@@ -39,11 +41,17 @@ public function create() {
     }
 
     public function edit(Service $servizio) {
-        return view('servizi.edit', compact('servizio'));
+        if ($servizio->user_id == Auth::user()->id) {
+            return view('servizi.edit', compact('servizio'));
+        } else {
+            return redirect()->route('home')->with('errorMessage', 'Non hai i permessi per modificare questo servizio.');
+        }
+        
     }
 
     public function update(ServiceRequest $request, Service $servizio) {
-      $servizio->update([
+       if ($servizio->user_id == Auth::user()->id) {
+    $servizio->update([
          'name' => $request->name,
          'description' => $request->description,
          'price' => $request->price,
@@ -52,12 +60,21 @@ public function create() {
             $servizio->img = $request->file('img')->store('images', 'public');
         }
        $servizio->save();
-
-
-       return redirect()->route('home')->with('successMessage','Servizio modificato con successo!');    }
+       return redirect()->route('home')->with('successMessage','Servizio modificato con successo!'); 
+           } else {
+            return redirect()->route('home')->with('errorMessage', 'Non hai i permessi per modificare questo servizio.');
+        }
+       
+       
+       }
 
        public function destroy(Service $servizio) {
-        $servizio->delete();
-        return redirect()->route('home')->with('successMessage','Servizio eliminato con successo!');    }
+        if ($servizio->user_id == Auth::user()->id) {
+       $servizio->delete();
+        return redirect()->route('home')->with('successMessage','Servizio eliminato con successo!');   
+         } else {
+            return redirect()->route('home')->with('errorMessage', 'Non hai i permessi per modificare questo servizio.');
+        }
+        }
 }
